feat(fmt): make_names_unique: simplify and promote to internal API
