+++ src/fmt/mol2.cpp
@@ -682,45 +682,6 @@ SubstructInfo resolve_substructs(const Molecule &mol) {
   return info;
 }
 
-struct NameMapEntry {
-  int first_idx;
-  int count = 1;
-  std::string safe_name = {};  // NOLINT(readability-redundant-member-init)
-};
-
-template <class NameType, class C, class NameFunc>
-std::vector<std::string> make_names_unique(const C &cont, NameFunc ith_name) {
-  std::vector<std::string> names;
-  names.reserve(cont.size());
-
-  absl::flat_hash_map<NameType, NameMapEntry> name_map;
-  name_map.reserve(cont.size());
-
-  for (int i = 0; i < cont.size(); ++i) {
-    NameType name = ith_name(i);
-
-    if (name.empty()) {
-      names.push_back({});
-      continue;
-    }
-
-    auto [it, first] = name_map.try_emplace(
-        std::move(name), NameMapEntry { static_cast<int>(names.size()) });
-    if (first) {
-      it->second.safe_name = internal::ascii_safe(it->first);
-      names.push_back(it->second.safe_name);
-      continue;
-    }
-
-    if (it->second.count == 1)
-      names[it->second.first_idx] = absl::StrCat(it->second.safe_name, "1");
-
-    names.push_back(absl::StrCat(it->second.safe_name, ++it->second.count));
-  }
-
-  return names;
-}
-
 std::string_view sybyl_subtype_hyb(constants::Hybridization hyb) {
   // NOLINTNEXTLINE(clang-diagnostic-switch-enum)
   switch (hyb) {
@@ -1099,16 +1060,15 @@ bool write_mol2(std::string &out, const Molecule &mol, int conf,
   int atom_id_width = width_of_size(mol.num_atoms()),
       bond_id_width = width_of_size(mol.num_bonds());
 
-  std::vector atom_names =
-      make_names_unique<std::string_view>(mol, [&mol](int i) {
-        auto atom = mol[i];
-        std::string_view name = atom.data().get_name();
-        return name.empty() ? atom.data().element_symbol() : name;
-      });
+  std::vector atom_names = make_names_unique(mol, [&mol](int i) {
+    auto atom = mol[i];
+    std::string_view name = atom.data().get_name();
+    return name.empty() ? atom.data().element_symbol() : name;
+  });
   std::vector atom_types = sybyl_atom_types(mol);
 
   const SubstructInfo subs_info = resolve_substructs(mol);
-  std::vector substruct_names = make_names_unique<std::string>(
+  std::vector substruct_names = make_names_unique(
       subs_info.root_of_sub, [&mol, &subs_info](int idx) -> std::string {
         int root = subs_info.root_of_sub[idx];
         if (root < 0)
